requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.9.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2023.7.22
channels==4.3.1
charset-normalizer==3.2.0
distlib==0.3.8
Django==4.2.9
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.13.1
flake8==3.9.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflate64==0.3.1
instaloader==4.10
keyboard==0.13.5
mccabe==0.6.1
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
multivolumefile==0.2.3
numpy==1.26.4
pandas==2.2.0
Pillow==10.0.1
platformdirs==4.2.0
psutil==5.9.5
psycopg2==2.9.9
py7zr==0.20.6
PyAutoGUI==0.9.54
pybcj==1.0.1
pycodestyle==2.7.0
pycryptodomex==3.19.0
pyflakes==2.3.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pyppmd==1.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-telegram-bot==20.4
pytweening==1.0.7
pytz==2023.3.post1
pyzstd==0.15.9
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
telepot==12.7
texttable==1.6.7
tzdata==2023.4
urllib3==2.0.4
virtualenv==20.25.0
yarl==1.9.2
